requirements.txt
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.0.0
PyPDF2==1.26.0
pytesseract==0.2.0
pyttsx3==2.7
requests==2.18.4
urllib3==1.22
Werkzeug==0.14.1
